+++ backend/apid/actions/checks.go
@@ -29,6 +29,8 @@ var checkConfigUpdateFields = []string{
 	"Timeout",
 	"Ttl",
 	"ProxyRequests",
+	"MetricFormat",
+	"MetricHandlers",
 }
 
 var (